0.34.0: Bump mongochangestream to pick up refresh token improvements
Also, use `assertEventually` to make tests more reliable when latency is bad:

I'm currently in Brazil and experiencing poor latency, which is making the tests
fail the way that they're written. I was able to get them passing by using
`assertEventually` to repeatedly check until the condition passes, failing once
a max timeout is reached.

(TODO: Remove this inlined version of `assertEventually` and pull it in from
mongochangestream-testing instead.)

Results:

   ? src/syncData.test.ts (2 tests) 19299ms
   ? syncCollection > initialScan should work 11639ms
   ? syncCollection > should process records via change stream 7659ms

 Test Files  1 passed (1)
      Tests  2 passed (2)
   Start at  13:58:43
   Duration  19.70s (transform 28ms, setup 0ms, collect 215ms, tests 19.30s, environment 0ms, prepare 38ms)